refactor modalEditor
