+++ admin/src/app/App.tsx
@@ -6,7 +6,7 @@ import { Header } from "widgets/Header/Header"
 import { RequireAuth } from "features/authByGoogle"
 import { StepsPage } from "pages/StepsPage"
 import { ServicesPage } from "pages/ServicesPage"
-import { ReviewsPage } from "pages/ReviewPage"
+import { ReviewsPage } from "pages/ReviewsPage"
 import { FAQPage } from "pages/FAQPage"
 import { OtherPage } from "pages/OtherPage"
 import { ContactsPage } from "pages/ContactPage/ui/ContactsPage"
