+++ admin/src/features/portfolioFilters/ui/AddNewItem/AddNewItem.tsx
@@ -1,7 +1,7 @@
 import { defaultNewFilter } from "../../const/const"
 import { IFilterItem, IFiltersData, IGlobalData } from "../../types/types"
 import { useState } from "react"
-import { ModalEditorWithTranslation } from "shared/components/ModalEditorWithTranslation/ModalEditorWithTranslation"
+import { ModalEditor } from "shared/components/ModalEditor/ModalEditor"
 import { updateFiltersData } from "shared/const/firebaseVariables"
 import { defaultLanguage } from "shared/const/languages"
 import { LanguageType } from "shared/types/types"
@@ -74,9 +74,10 @@ export function AddNewItem({
 
     return (
         <>
-            <ModalEditorWithTranslation
+            <ModalEditor
+                withTranslation
                 isOpen={isOpen}
-                onClose={() => setIsOpen(true)}
+                onClose={() => setIsOpen(false)}
                 currentLanguage={currentLanguage}
                 onChangeLanguage={onChangeLanguage}
                 onSaveClick={saveClickHandler}
@@ -88,7 +89,7 @@ export function AddNewItem({
                         setNewAllItem(prev => ({ ...prev, [currentLanguage]: value }))
                     }
                 />
-            </ModalEditorWithTranslation>
+            </ModalEditor>
             <button className={styles.btn} onClick={() => setIsOpen(true)}>
                 Add New Item
             </button>
