+++ admin/src/pages/ContactPage/ui/EditContact/EditContact.tsx
@@ -1,36 +1,28 @@
 import { useState } from "react"
-import { ModalEditorWithTranslation } from "shared/components/ModalEditorWithTranslation/ModalEditorWithTranslation"
-import { LanguageType } from "shared/types/types"
+import { ModalEditor } from "shared/components/ModalEditor/ModalEditor"
 import styles from "./EditContact.module.scss"
 import { Input } from "shared/ui/Input/Input"
 
 export function EditContact({ label, contact }: { label: string; contact: string }) {
     const [info, setInfo] = useState("")
     const [isOpen, setIsOpen] = useState(false)
-    const [currentLanguage, setCurrentLanguage] = useState<LanguageType>("en")
 
     function onClose() {
         setIsOpen(false)
     }
 
-    function onChangeLanguage(lang: LanguageType) {
-        setCurrentLanguage(lang)
-    }
-
     return (
         <>
-            <ModalEditorWithTranslation
+            <ModalEditor
                 isOpen={isOpen}
                 onClose={onClose}
-                onChangeLanguage={onChangeLanguage}
-                currentLanguage={currentLanguage}
                 onSaveClick={() => null}
                 onDiscardClick={() => null}
             >
                 <div className={styles.container}>
                     <Input label={label} value={contact} onChange={value => setInfo(value)} />
                 </div>
-            </ModalEditorWithTranslation>
+            </ModalEditor>
             <button onClick={() => setIsOpen(true)}>Edit</button>
         </>
     )
