+++ admin/src/pages/ContactPage/ui/EditSocialMedia/EditSocialMedia.tsx
@@ -1,31 +1,23 @@
 import { useState } from "react"
-import { ModalEditorWithTranslation } from "shared/components/ModalEditorWithTranslation/ModalEditorWithTranslation"
+import { ModalEditor } from "shared/components/ModalEditor/ModalEditor"
 import { ModalImage } from "shared/components/ModalImage/ModalImage"
 import { Input } from "shared/ui/Input/Input"
 import styles from "./EditSocialMedia.module.scss"
-import { LanguageType } from "shared/types/types"
 import { ISocialMedia } from "../../types/type"
 
 export function EditSocialMedia({ social }: { social: ISocialMedia }) {
     const [data, setData] = useState<ISocialMedia>({ icon: "", link: "" })
     const [isOpen, setIsOpen] = useState(false)
-    const [currentLanguage, setCurrentLanguage] = useState<LanguageType>("en")
 
     function onClose() {
         setIsOpen(false)
     }
 
-    function onChangeLanguage(lang: LanguageType) {
-        setCurrentLanguage(lang)
-    }
-
     return (
         <>
-            <ModalEditorWithTranslation
+            <ModalEditor
                 isOpen={isOpen}
                 onClose={onClose}
-                onChangeLanguage={onChangeLanguage}
-                currentLanguage={currentLanguage}
                 onSaveClick={() => null}
                 onDiscardClick={() => null}
             >
@@ -41,7 +33,7 @@ export function EditSocialMedia({ social }: { social: ISocialMedia }) {
                         onChange={value => setData(prev => ({ ...prev, link: value }))}
                     />
                 </div>
-            </ModalEditorWithTranslation>
+            </ModalEditor>
             <button onClick={() => setIsOpen(true)}>Edit</button>
         </>
     )
