+++ admin/src/pages/ReviewsPage/ui/DeleteParagraph/DeleteParagraph.tsx
