+++ admin/src/pages/ReviewsPage/ui/EditParagraph/EditParagraph.module.scss
