+++ admin/src/pages/ServicesPage/ui/AddServiceModal/AddServiceModal.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react"
-import { ModalEditorWithTranslation } from "shared/components/ModalEditorWithTranslation/ModalEditorWithTranslation"
+import { ModalEditor } from "shared/components/ModalEditor/ModalEditor"
 import { Input } from "shared/ui/Input/Input"
 import { Textarea } from "shared/ui/Textarea/Textarea"
 import styles from "./AddServiceModal.module.scss"
 import { LanguageType } from "shared/types/types"
-import { INewAllData, IPreviewImage, IServiceData, ITranslatedServiceData } from "../../types/types"
+import { INewAllData, IPreviewImage, ITranslatedServiceData } from "../../types/types"
 import { ImagesList } from "../ImagesList/ImagesList"
 import { EditImage } from "features/editImage"
 import { allLanguages, defaultLanguage } from "shared/const/languages"
@@ -112,7 +112,8 @@ export function AddServiceModal({
     return (
         <>
             <LoadingModal isLoading={isLoading} />
-            <ModalEditorWithTranslation
+            <ModalEditor
+                withTranslation
                 isOpen={isOpen}
                 onClose={onClose}
                 onChangeLanguage={onChangeLanguage}
@@ -150,7 +151,7 @@ export function AddServiceModal({
                         }
                     />
                 </div>
-            </ModalEditorWithTranslation>
+            </ModalEditor>
             <button onClick={() => setIsOpen(true)}>Add New</button>
         </>
     )
