+++ admin/src/shared/components/ModalEditor/ModalEditor.module.scss
