+++ admin/src/widgets/Sidebar/ui/Sidebar.module.scss
@@ -35,6 +35,7 @@
         font-size: 18px;
         box-shadow: 1px 1px 0 0 var(--color-moderate-accent);
         text-decoration: none;
+        transition: ease 0.3s;
 
         &.currentLink {
             display: flex;
@@ -49,13 +50,13 @@
         }
     }
     .linkBlock:hover .link {
-        transition: ease 0.3s;
         transform: translateX(10%);
         cursor: pointer;
         color: var(--color-white);
         background-color: var(--color-moderate-accent);
         border-right: 1px solid var(--color-base-accent);
         box-shadow: 0 1px 0 0 var(--color-base-accent);
+        border-radius: 10px;
     }
 
     .svg path {
